+++ src/NadekoBot/Services/CommandHandler.cs
@@ -151,9 +151,9 @@ namespace NadekoBot.Services
         }
 
         private bool IsBlacklisted(IGuild guild, IUserMessage usrMsg) =>
-            (guild != null && BlacklistCommands.BlacklistedGuilds.Contains(guild.Id)) ||
-            BlacklistCommands.BlacklistedChannels.Contains(usrMsg.Channel.Id) ||
-            BlacklistCommands.BlacklistedUsers.Contains(usrMsg.Author.Id);
+            (guild != null && BlacklistCommands.BlacklistedGuilds.Contains((long) guild.Id)) ||
+            BlacklistCommands.BlacklistedChannels.Contains((long) usrMsg.Channel.Id) ||
+            BlacklistCommands.BlacklistedUsers.Contains((long) usrMsg.Author.Id);
 
         private const float _oneThousandth = 1.0f / 1000;
         private Task LogSuccessfulExecution(IUserMessage usrMsg, ExecuteCommandResult exec, ITextChannel channel, int exec1, int exec2, int exec3, int total)
